slow analyzer

+++ modules/msp/apm/service/components/transaction-cache-error/metric_table_filter/provider.go
@@ -56,8 +56,8 @@ type State struct {
 
 type FrontendConditions struct {
 	Duration []struct {
-		Timer int64  `json:"timer,omitempty"`
-		Unit  string `json:"unit,omitempty"`
+		Timer float64 `json:"timer,omitempty"`
+		Unit  string  `json:"unit,omitempty"`
 	} `json:"duration,omitempty"`
 }
 
@@ -66,11 +66,11 @@ func (f *FrontendConditions) convertToTransactionFilter() error_transaction.Erro
 	if len(f.Duration) != 2 {
 		return errorTransactionFilter
 	}
-	d, err := time.ParseDuration(fmt.Sprintf("%d%s", f.Duration[0].Timer, f.Duration[0].Unit))
+	d, err := time.ParseDuration(fmt.Sprintf("%f%s", f.Duration[0].Timer, f.Duration[0].Unit))
 	if err == nil {
 		errorTransactionFilter.MinDuration = float64(d)
 	}
-	d, err = time.ParseDuration(fmt.Sprintf("%d%s", f.Duration[1].Timer, f.Duration[1].Unit))
+	d, err = time.ParseDuration(fmt.Sprintf("%f%s", f.Duration[1].Timer, f.Duration[1].Unit))
 	if err == nil {
 		errorTransactionFilter.MaxDuration = float64(d)
 	}
