+++ modules/msp/apm/service/components/transaction-db-slow/metric_table_filter/provider.go
@@ -56,8 +56,8 @@ type State struct {
 
 type FrontendConditions struct {
 	Duration []struct {
-		Timer int64  `json:"timer,omitempty"`
-		Unit  string `json:"unit,omitempty"`
+		Timer float64 `json:"timer,omitempty"`
+		Unit  string  `json:"unit,omitempty"`
 	} `json:"duration,omitempty"`
 }
 
@@ -66,7 +66,7 @@ func (f *FrontendConditions) convertToTransactionFilter() slow_transaction.SlowT
 	if len(f.Duration) != 2 {
 		return slowTransactionFilter
 	}
-	d, err := time.ParseDuration(fmt.Sprintf("%d%s", f.Duration[0].Timer, f.Duration[0].Unit))
+	d, err := time.ParseDuration(fmt.Sprintf("%f%s", f.Duration[0].Timer, f.Duration[0].Unit))
 	if err == nil {
 		slowTransactionFilter.MinDuration = float64(d)
 	}
